+++ src/cli/import.rs
@@ -18,7 +18,8 @@ use miden_objects::{
 use miden_tx::auth::TransactionAuthenticator;
 use tracing::info;
 
-use super::Parser;
+use super::{utils::load_config_file, Parser};
+use crate::cli::account::maybe_set_default_account;
 
 #[derive(Debug, Parser, Clone)]
 #[clap(about = "Import client objects such as accounts and notes")]
@@ -37,6 +38,7 @@ impl ImportCmd {
         mut client: Client<N, R, S, A>,
     ) -> Result<(), String> {
         validate_paths(&self.filenames)?;
+        let (mut current_config, _) = load_config_file()?;
         for filename in &self.filenames {
             let note_id = import_note(&mut client, filename.clone(), !self.no_verify).await;
             if note_id.is_ok() {
@@ -46,6 +48,10 @@ impl ImportCmd {
             let account_id = import_account(&mut client, filename)
                 .map_err(|_| format!("Failed to parse file {}", filename.to_string_lossy()))?;
             println!("Succesfully imported account {}", account_id);
+
+            if account_id.is_regular_account() {
+                maybe_set_default_account(&mut current_config, account_id)?;
+            }
         }
         Ok(())
     }