+++ src/client/sync.rs
@@ -444,7 +444,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
                     note_recipient,
                 );
 
-                let input_note = InputNote::new(note, note_inclusion_proof);
+                let input_note = InputNote::authenticated(note, note_inclusion_proof);
 
                 tracked_input_notes.push(input_note);
             }
@@ -515,7 +515,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
                     )
                     .map_err(ClientError::NoteError)?;
 
-                    return_notes.push(InputNote::new(note, note_inclusion_proof))
+                    return_notes.push(InputNote::authenticated(note, note_inclusion_proof))
                 },
             }
         }
@@ -668,10 +668,9 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
             idx = idx.sibling();
             // Rightmost index is always the biggest value, so if the path contains any node
             // past it, we can discard it for our version of the forest
-            if idx > rightmost_index {
-                continue;
+            if idx <= rightmost_index {
+                path_nodes.push((idx, node));
             }
-            path_nodes.push((idx, node));
             idx = idx.parent();
         }