+++ tests/integration/main.rs
@@ -15,6 +15,7 @@ use miden_tx::TransactionExecutorError;
 mod common;
 use common::*;
 
+mod cli_tests;
 mod custom_transactions_tests;
 mod onchain_tests;
 mod swap_transactions_tests;